backend/pyproject.toml
[project]
name = "lichess-backend"
version = "0.1.0"
description = "Lichess stats backend"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "ruff>=0.14.11",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_awaitable = true
strict_optional = true
ignore_missing_imports = true
